+++ app/src/main/java/org/sic4change/nut4health/data/entities/Contract.java
@@ -30,11 +30,11 @@ public class Contract {
 
     @NonNull
     @ColumnInfo(name = DataContractNames.COL_LATITUDE)
-    private float latitude;
+    private double latitude;
 
     @NonNull
     @ColumnInfo(name = DataContractNames.COL_LONGITUDE)
-    private float longitude;
+    private double longitude;
 
     @NonNull
     @ColumnInfo(name = DataContractNames.COL_SCREENER)
@@ -88,27 +88,27 @@ public class Contract {
     }
 
     @Ignore
-    public Contract(@NonNull String photo, float latitude, float longitude, @NonNull String screener) {
+    public Contract(@NonNull String photo, double latitude, double longitude, @NonNull String screener) {
         this("", photo, latitude, longitude, screener, "", "", "",
                 "", "", Status.INIT.name(), "", "",0,  0, "");
     }
 
     @Ignore
-    public Contract(@NonNull String photo, float latitude, float longitude, @NonNull String screener,
+    public Contract(@NonNull String photo, double latitude, double longitude, @NonNull String screener,
                     String childName, String childSurname, String childAddress, String fingerprint, String status, String creationDate,
                     int percentage) {
         this("", photo, latitude, longitude, screener, "", childName, childSurname,
                 childAddress, fingerprint, status, "", creationDate, Nut4HealthTimeUtil.convertCreationDateToTimeMilis(creationDate), percentage, "");
     }
     @Ignore
-    public Contract(@NonNull String id, @NonNull String photo, float latitude, float longitude,
+    public Contract(@NonNull String id, @NonNull String photo, double latitude, double longitude,
                     @NonNull String screener, String childName, String childSurname,
                     String childAddress, String fingerprint, String status, String creationDate, int percentage) {
         this(id, photo, latitude, longitude, screener, "", childName, childSurname,
                 childAddress, fingerprint, status, "", creationDate, Nut4HealthTimeUtil.convertCreationDateToTimeMilis(creationDate), percentage, "");
     }
 
-    public Contract(@NonNull String id, @NonNull String photo, float latitude, float longitude, @NonNull String screener,
+    public Contract(@NonNull String id, @NonNull String photo, double latitude, double longitude, @NonNull String screener,
                     String medical, String childName, String childSurname, String childAddress, String fingerprint,
                     String status, String diagnosis, String creationDate, long creationDateMiliseconds, int percentage, String medicalDate) {
         this.id = id;
@@ -148,19 +148,19 @@ public class Contract {
         this.photo = photo;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
